Add Sky Scene Add-on to Memory Management List
SkyScene Add-on is a Magisk/KSU module that is a fork (i.e., open source because the original version is also open source) that modifies most of the memory management subsystems with the sole purpose of improving memory management and battery saving, pushing the limits of data storage capacity and the ability to retain it in memory. It contains customizable parameters and even features that only OEMs have.